+++ spec/features/model_spec.rb
@@ -1,5 +1,27 @@
 # frozen_string_literal: true
 
+class TestValidator1 < ActiveModel::Validator
+  def validate(record)
+    options[:attributes].each do |attribute_name|
+      attribute_value = record.send(attribute_name)
+      if attribute_value == :bad
+        record.errors.add(attribute_name, "can't be bad")
+      end
+    end
+  end
+end
+
+class TestValidator2 < ActiveModel::Validator
+  def validate(record)
+    options[:fields].each do |attribute_name|
+      attribute_value = record.send(attribute_name)
+      if attribute_value == :bad
+        record.errors.add(attribute_name, "can't be bad")
+      end
+    end
+  end
+end
+
 RSpec.describe 'DecoLite::Model features', type: :features do
   context 'when defining attributes' do
     context 'when loaded fields do not conflict with existing attributes' do
@@ -73,9 +95,12 @@ RSpec.describe 'DecoLite::Model features', type: :features do
       subject do
         Class.new(DecoLite::Model) do
           validates :field1, :field2, :field3, presence: true
+          validates_with TestValidator1, attributes: %i[field5 field6]
+          validates_with TestValidator2, fields: %i[field7 field8]
+          validates_with TestValidator1, attributes: %i[field9]
 
           def required_fields
-            %i[field4]
+            %i[field4 field9]
           end
         end.new(hash: hash, options: options)
       end
@@ -84,16 +109,38 @@ RSpec.describe 'DecoLite::Model features', type: :features do
         subject.validate
       end
 
-      let(:hash) { { field1: :value1, field2: :value2, field3: nil } }
+      let(:hash) do
+        {
+          field1: :value1,
+          field2: :value2,
+          field3: nil,
+          field5: :bad,
+          field6: :good,
+          field7: :bad,
+          field8: :good
+        }
+      end
 
-      it 'does not overwrite the loaded field values' do
+      it 'maintains the loaded field values' do
         expect(subject.field1).to eq hash[:field1]
         expect(subject.field2).to eq hash[:field2]
         expect(subject.field3).to eq hash[:field3]
+        # field4 is missing
+        expect(subject.field5).to eq hash[:field5]
+        expect(subject.field6).to eq hash[:field6]
+        expect(subject.field7).to eq hash[:field7]
+        expect(subject.field8).to eq hash[:field8]
+        # field9 is missing
       end
 
       it 'validates correctly' do
-        expected_errors = ["Field3 can't be blank", "Field4 field is missing"]
+        expected_errors = [
+          "Field3 can't be blank",
+          "Field4 field is missing",
+          "Field5 can't be bad",
+          "Field7 can't be bad",
+          "Field9 field is missing",
+        ]
         expect(subject.errors.full_messages).to match_array expected_errors
       end
     end